fix for AmazonClientException: ConnectionPoolTimeoutException
S3Object opens a connection for each object.
These connections are not liberated even if the object is garbage collected,
so it is needed to execute either use Try-with-resources (preferred)
or manually call object.close() in order to liberate the connection to the pool.
Failing to do so will result in AmazonClientException: ConnectionPoolTimeoutException.

More information can be found here:
- http://stackoverflow.com/questions/17782937/connectionpooltimeoutexception-when-iterating-objects-in-s3
- https://forums.aws.amazon.com/message.jspa?messageID=471149
- https://java.awsblog.com/post/Tx22D24MCRWFR9U/Closeable-S3Objects
- http://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/s3/model/S3Object.html